+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 Version 172:
 
 * Tidy up websocket stream javadocs
+* Fix move-only arguments in bind_handler
 
 --------------------------------------------------------------------------------